backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.101.0
pymupdf==1.26.4
PyPDF2==3.0.1
pydantic==2.5.0
aiofiles==23.2.1
tiktoken==0.11.0
pyyaml==6.0.2
cors
fastapi-cors
